Common signer and coordinator functionality extracted to Participant
Participant struct is a base [FROST] participant, no matter the type:
signer, coordinator, or both. The logic is used from the Signer and soon
will be used from the Coordinator as well (see next commits).
