--- frost/frost.go
@@ -1,1 +1,0 @@
-package frost
